Tidy up transition class types
