+++ dsl/src/main/java/tech/pegasys/peeps/node/genesis/transitions/BesuTransitions.java
@@ -10,6 +10,23 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package tech.pegasys.peeps.node.genesis.qbft;
+package tech.pegasys.peeps.node.genesis.transitions;
 
-public interface Transition {}
+import java.util.ArrayList;
+import java.util.List;
+
+import com.fasterxml.jackson.annotation.JsonGetter;
+import tech.pegasys.peeps.node.genesis.transitions.Transition;
+
+public class BesuTransitions {
+  private final List<Transition> transitions = new ArrayList<>();
+
+  @JsonGetter("qbft")
+  public List<Transition> getTransitions() {
+    return transitions;
+  }
+
+  public void add(final Transition transition) {
+    this.transitions.add(transition);
+  }
+}
